+++ src/collective/metadataversion/config.py
@@ -5,3 +5,7 @@
 # code rather than simply by catalog.xml
 VERSION_KEY = 'metadata_version'
 FULL_VERSION_KEY = 'collective.metadataversion.interfaces.IMetadataSettings.metadata_version'
+
+# Which indexes to we refresh for "metadata-only" reindexing?
+FULL_IDXS_KEY = 'collective.metadataversion.interfaces.IMetadataSettings.default_idxs'
+DEFAULT_IDXS = ['getId']