+++ CHANGES.rst
@@ -2,6 +2,23 @@ Changelog
 =========
 
 
+0.2.0 (2021-11-17)
+------------------
+
+Bugfixes:
+
+- Register the ``metadata_version`` indexer,
+  and
+- attach it to IContentish
+
+New Features:
+
+- New setting ``default_idxs`` to adjust the (cheap!) indexes
+  to be refreshed for "metadata-only" updates
+
+[tobiasherp]
+
+
 0.1.0 (2021-09-21)
 ------------------